fixed silly bug with zyes_to_all.sh...

+++ zscripts/zyes_to_all.sh
@@ -4,16 +4,23 @@ then
   exit 1
 fi
 
-dir1=$IMARUN/output
-dir2=$IMARUN/output_to_javier
-dir3=$IMARUN/debug_output
+dir1=$IMAWRK/output
+dir2=$IMAWRK/output_to_javier
+dir3=$IMAWRK/debug_output
 
 zreplace(){
   local dir=${1}
-  cd $dir
-  pwd
-  rm -f z*.sh
-  cp $IMASMS/zscripts/z*.sh .
+  local myerr="$(cd $dir 2>&1)"
+  if [[ $myerr = *"No such file or directory"* ]]
+  then
+    echo $myerr
+    echo 'copying failed, directory skipped...'
+  else
+    cd $dir
+    pwd
+    rm -f z*.sh
+    cp $IMASMS/zscripts/z*.sh .
+  fi
 }
 
 echo 'updating all relevant output directories with zscripts from $IMASMS...'
